Movement + SpriteSheet animations + level art
Python:
- created a python script to create spritesheets from image sequences
- created a python script to create a multispritesheet containing multiple other spritesheets
in the ouput log, I print a mapping from the spriteSheet name to where the spritesheet starts and ends in the multispritesheet

C++:
- Created a Spritesheet class
- Created a MultiSpriteSheet class that inherits Spritesheet
- Fixed Dashing velocity being super large
- Correct Animation plays per state
- Reworked how the data for each state is stored,
All the variables for how madeline should move in a certain state are now saved in
a MovementParameters vector. Each state has a pointer to which MovementParameters it should use in x and y.
- Animation is flipped based on movement dir or when grabbing a wall left or right
- Implemented basic system that maps a certain pixel value in the level image to the correct texture/tile it should draw. (currently hardcoded for 1 level)